test: moved around tests
