+++ tests/test_building_options.py
@@ -1,7 +1,9 @@
+import enum
 import pytest
 import gqlrequests
-import enum
+
 from typing import List
+
 
 class EveryType(gqlrequests.QueryBuilder):
     id: int
@@ -9,8 +11,8 @@ class EveryType(gqlrequests.QueryBuilder):
     money: float
     name: str
     company: bool
-
-def test_primitive_types():
+    
+def test_class_with_only_primitives_builds_correctly():
     correct_string = """
 {
     id
@@ -21,8 +23,8 @@ def test_primitive_types():
 }
 """[1:]
     assert EveryType().build() == correct_string
-
-def test_indent():
+    
+def test_setting_indent_builds_correctly():
     correct_string = """
 {
   id
@@ -34,17 +36,7 @@ def test_indent():
 """[1:]
     assert EveryType().build(indent_size=2) == correct_string
 
-
-def test_query_selection():
-    correct_string = """
-{
-    id
-    age
-}
-"""[1:]
-    assert EveryType(fields=["id", "age"]).build() == correct_string
-
-def test_start_indent():
+def test_builder_with_start_indents_set_builds_correctly():
     # Note that the first bracket is not indented. This is intentional because
     # it is meant to be used in as a field value in a query, where the first
     # bracket is only 1 whitespace away from the field name
@@ -59,7 +51,7 @@ def test_start_indent():
 """[1:]
     assert EveryType().build(indent_size=2, start_indents=2) == correct_string
 
-def test_selecting_no_fields_and_converting_to_string_raises_value_error():
+def test_selecting_no_fields_and_building_raises_value_error():
     with pytest.raises(ValueError) as e:
         EveryType(fields=[]).build()
 
@@ -71,24 +63,7 @@ class NestedType(gqlrequests.QueryBuilder):
     age: int
     something: EveryType
 
-
-def test_nested_indent():
-    correct_string = """
-{
-  id
-  age
-  something {
-    id
-    age
-    money
-    name
-    company
-  }
-}
-"""[1:]
-    assert NestedType().build(indent_size=2) == correct_string
-
-def test_nested_types():
+def test_nested_type_builds_correctly():
     correct_string = """
 {
     id
@@ -105,6 +80,23 @@ def test_nested_types():
     assert NestedType().build() == correct_string
 
 
+def test_nested_indent_builds_correctly():
+    correct_string = """
+{
+  id
+  age
+  something {
+    id
+    age
+    money
+    name
+    company
+  }
+}
+"""[1:]
+    assert NestedType().build(indent_size=2) == correct_string
+
+
 class ListedType(gqlrequests.QueryBuilder):
     id: int
     names: List[str]
@@ -112,7 +104,7 @@ class ListedType(gqlrequests.QueryBuilder):
 
 
 @pytest.mark.skip(reason="Not implemented yet")
-def test_listed_types():
+def test_class_with_list_type_as_field_builds_correctly():
     correct_string = """
 {
     id
@@ -128,42 +120,6 @@ def test_listed_types():
 """
     assert ListedType().build() == correct_string
 
-
-class SomeClass(gqlrequests.QueryBuilder):
-    pass
-
-
-class InvalidType(gqlrequests.QueryBuilder):
-    invalidProperty: SomeClass
-
-@pytest.mark.skip(reason="Not implemented yet")
-def test_invalid_type_throws_value_error():
-    with pytest.raises(ValueError) as e:
-        InvalidType()
-
-    # Ensure the error message contains the name of the invalid property 
-    # and the name of the class - this is to help the user debug the issue
-    assert "invalidProperty" in str(e.value) and "InvalidType" in str(e.value)
-
-
-class DatatypeWithKeywordAsProperty(gqlrequests.QueryBuilder):
-    type: str  # This is okay, because type is a reserved *function*, not keyword
-    from_: str
-    as_: int
-
-
-@pytest.mark.skip(reason="Not implemented yet")
-def test_keyword_as_property_gets_stripped_for_underscores():
-    correct_string = """
-{
-    type
-    from
-    as
-}
-"""
-    assert DatatypeWithKeywordAsProperty().build() == correct_string
-
-
 class SomeEnumClass(enum.Enum):
     ONE = 1
     TWO = 2
@@ -174,10 +130,28 @@ class ClassWithEnumeratedFieldType(gqlrequests.QueryBuilder):
     enumerated: SomeEnumClass
 
 
-def test_enumerated_field_type():
+def test_class_with_enum_field_type_builds_correctly():
     correct_string = """
 {
     enumerated
 }
 """[1:]
     assert ClassWithEnumeratedFieldType().build() == correct_string
+
+
+class DatatypeWithKeywordAsProperty(gqlrequests.QueryBuilder):
+    type: str  # This is okay, because type is a reserved *function*, not keyword
+    from_: str
+    as_: int
+
+
+@pytest.mark.skip(reason="Not implemented yet")
+def test_keywords_as_field_name_gets_stripped_for_underscores():
+    correct_string = """
+{
+    type
+    from
+    as
+}
+"""
+    assert DatatypeWithKeywordAsProperty().build() == correct_string
