+++ tests/test_function_queries.py
