+++ tests/test_query_creator_helpers.py
